refactor: 음식점 정보 모달 컨트롤러 분리
